初始-样例

+++ liv-auth/src/main/java/com/liv/auth/controller/MenuController.java
@@ -5,14 +5,19 @@ import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.liv.api.auth.dao.MenuMapper;
 import com.liv.api.auth.dao.datamodel.Menu;
+import com.liv.api.auth.dao.datamodel.MenuQuery;
+import com.liv.api.auth.domainmodel.MenuDO;
 import com.liv.api.auth.service.MenuService;
 import com.liv.api.base.annotation.ValidResult;
 import com.liv.api.base.base.BaseController;
 import com.liv.api.base.base.DataBody;
 import com.liv.api.base.base.ResultBody;
+import com.liv.auth.dao.datamodel.App;
+import com.liv.auth.dao.datamodel.AppQuery;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiImplicitParam;
 import io.swagger.annotations.ApiOperation;
+import org.apache.commons.collections4.ListUtils;
 import org.apache.shiro.SecurityUtils;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
@@ -38,19 +43,24 @@ public class MenuController extends BaseController<MenuMapper, Menu, MenuService
         return ResultBody.success(service.getById(MenuId));
     }
 
+//    @ApiOperation(value = "分页查询菜单列表", notes="分页查询菜单列表")
+//    @PostMapping(value="/pagelist")
+//    public DataBody pagelist(@RequestBody MenuQuery query) throws Exception {
+//        QueryWrapper<Menu> qw = query.getQueryWrapper();
+//        IPage<MenuDO>  pageList = service.findPageList(query.getPage(),qw);
+//        return DataBody.success(pageList);
+//    }
+
     @ApiOperation(value = "查询菜单列表", notes="查询菜单列表")
-    @GetMapping(value="/list")
-    public List<Menu> list() throws Exception {
-        return service.list();
+    @PostMapping(value="/list")
+    public DataBody list(@RequestBody Menu menu) throws Exception {
+        return DataBody.success(service.getTreeList(menu));
     }
 
     @ApiOperation(value = "分页查询菜单列表", notes="分页查询菜单列表")
-    @GetMapping(value="/pagelist")
-    public DataBody pagelist(int current, int size) throws Exception {
-        QueryWrapper<Menu> wrapper = new QueryWrapper();
-        Page<Menu> page = new Page<>(current,size);
-        IPage<Menu> pageList = service.page(page,wrapper);
-        return DataBody.success(pageList);
+    @GetMapping(value="/queryByParentId")
+    public DataBody queryByParentId(Long parentId) throws Exception {
+        return DataBody.success(service.findByParentId(parentId));
     }
 
     @ApiOperation(value = "新增菜单", notes="新增菜单")
@@ -71,9 +81,13 @@ public class MenuController extends BaseController<MenuMapper, Menu, MenuService
 
     @ApiOperation(value = "删除菜单", notes="删除菜单,根据主键id删除")
     @DeleteMapping(value="/remove/{id}")
-    public ResultBody delete(@PathVariable("id") Long id){
-        service.removeById(id);
-        return ResultBody.success("删除成功");
+    public ResultBody delete(@PathVariable("id") Long id) throws Exception {
+        List<MenuDO> list = service.findByParentId(id);
+        if(list!=null&&list.size()>0){
+            return ResultBody.error("删除失败，请先删除下级菜单");
+        }
+        boolean flag = service.removeById(id);
+        return ResultBody.success(flag?"删除成功":"删除失败，未删除任何数据");
     }
 
     @ApiOperation(value = "批量删除菜单", notes="删除菜单,根据主键id删除")
